Unify simulation logic between the two modes
