+++ public/js/advanced-mode.js
@@ -17,17 +17,12 @@ import {
     addMetricsButtons,
     addDimensionsButtons,
     getEpsilonFromDom,
-    getDailyValue,
     getBatchingFrequencyFromDom,
     getAllDimensionSizes,
-    displaySimulationResults_advancedMode,
-    displaySimulationResults_simpleMode,
     displaySimulationResults_unified,
     getMetricsArrayFromDom,
-    createSimulationDiv,
     getDimensionsArrayFromDom,
     getIsKeyStrategyGranularFromDom,
-    getStrategiesKeyCombinations,
     getAllDimensionNamesFromDom,
     displayMetrics,
     displayDimensions,
@@ -37,36 +32,21 @@ import {
     addDimension,
     removeDimension,
     getIsUseScalingFromDom,
-    getKeyCombinationString,
-    validateInputsBeforeSimulation,
-    getBudgetValueForMetricIdFromDom,
     displayBudgetSplit,
-    getIsPercentageBudgetSplitFromDom,
     loadPython,
-    getZeroConversionsPercentageFromDom,
     getEventCount,
-    getKeyStrategy,
-    getKeyStrategyFromDom
 } from './dom'
-import { generateSimulationId, 
-         generateSimulationTitle,
-         tempSaveTable, 
-         downloadAll } from './utils.misc'
+import {
+    tempSaveTable,
+    downloadAll
+} from './utils.misc'
 import {
     CONTRIBUTION_BUDGET,
     DEFAULT_MEASUREMENT_GOALS,
     DEFAULT_DIMENSIONS,
 } from './config'
 
-import {
-    getRandomLaplacianNoise,
-    getScalingFactorForMetric,
-    calculateNoisePercentage,
-    generateKeyCombinationArray,
-    generateSummaryValue,
-    calculateAverageNoisePercentageRaw,
-    getNoise_Rmsre,
-} from './utils.noise'
+import { triggerSimulation } from './main'
 
 let allSimulationDataTables_advancedMode = {}
 
@@ -93,99 +73,9 @@ export function initializeDisplay_advancedMode() {
     addDimensionsButtons()
     displayBudgetSplit()
 }
-function simulatePerMetric(
-    keyCombinations,
-    metric,
-    epsilon,
-    contributionBudget,
-    isUseScaling,
-    batchingFrequency,
-    dailyCount,
-    simulationNo
-) {
-    const value = getBudgetValueForMetricIdFromDom(metric.id)
-    const isPercentage = getIsPercentageBudgetSplitFromDom()
-    const scalingFactor = isUseScaling
-        ? getScalingFactorForMetric(
-              metric,
-              value,
-              isPercentage,
-              contributionBudget
-          )
-        : 1
-    const keyCombinationString = getKeyCombinationString(keyCombinations.names)
-    console.log("generate string")
-    console.log(keyCombinationString)
-   
-    const report = []
-
-    var noisePercentageSum = 0
-    for (let i = 0; i < keyCombinations.combinations.length; i++) {
-        const noise = getRandomLaplacianNoise(contributionBudget, epsilon)
-
-        const randCount = generateSummaryValue(
-            metric,
-            i,
-            dailyCount,
-            batchingFrequency,
-            getZeroConversionsPercentageFromDom()
-        )
-
-        const noiseValueAPE = calculateNoisePercentage(
-            noise,
-            // Noiseless summary value
-            randCount * scalingFactor
-        )
-        noisePercentageSum += noiseValueAPE
-
-        const summaryValue_scaled_noisy = randCount * scalingFactor + noise
-
-        report.push({
-            key: keyCombinations.combinations[i],
-            summaryValue: randCount,
-            summaryValue_scaled_noiseless: randCount * scalingFactor,
-            summaryValue_scaled_noisy: summaryValue_scaled_noisy,
-            noise: noise,
-            noise_ape_individual: noiseValueAPE,
-        })
-    }
-
-    const allSummaryValuesPreNoise = Object.values(report).map(
-        (v) => v.summaryValue_scaled_noiseless
-    )
-    const allSummaryValuesPostNoise = Object.values(report).map(
-        (v) => v.summaryValue_scaled_noisy
-    )
-
-    const noise_ape = calculateAverageNoisePercentageRaw(
-        noisePercentageSum,
-        keyCombinations.combinations.length
-    )
-
-    const noise_rmsre = getNoise_Rmsre(
-        allSummaryValuesPostNoise,
-        allSummaryValuesPreNoise,
-        scalingFactor
-    )
-
-    const simulationReport = {
-        data: report,
-        noiseMetrics: {
-            noise_ape_percent: noise_ape,
-            noise_rmsre: noise_rmsre 
-        },
-        scalingFactor: scalingFactor,
-        measurementGoal: metric.name,
-        dimensionsString: keyCombinationString,
-        simulationNo: simulationNo
-       
-    }
-
-      return simulationReport
-}
 
 export function simulateAndDisplayResultsAdvancedMode() {
-    triggerSimulation(
+    const simulation = triggerSimulation(
         getMetricsArrayFromDom(),
         getDimensionsArrayFromDom(),
         getAllDimensionNamesFromDom(),
@@ -197,6 +87,8 @@ export function simulateAndDisplayResultsAdvancedMode() {
         getBatchingFrequencyFromDom(),
         getEventCount()
     )
+
+    displaySimulationResults_unified(simulation, "advanced")
 }
 
 export function resetMetrics() {
@@ -209,109 +101,6 @@ export function resetBudgetSplit() {
 
 export function resetDimensions() {
     displayDimensions(DEFAULT_DIMENSIONS)
-}
-
-// generate dataset
-function triggerSimulation(
-    metrics,
-    dimensions,
-    dimensionNames,
-    dimensionSizes,
-    epsilon,
-    contributionBudget,
-    isUseScaling,
-    isGranular,
-    batchingFrequency,
-    dailyConversionCount
-) {
-    // Validate inputs are correct
-    if (!validateInputsBeforeSimulation(metrics, dimensions, isGranular)) return
-
-    // declare array containing possible combinations for keys
-    var r = []
-    var keyCombList = []
-
-    // logic for generating one dataset with all keys - string parameter 'all' is used
-    if (isGranular) {
-        var keyComb = generateKeyCombinationArray(dimensionSizes)
-        keyCombList.push({
-            names: dimensionNames,
-            combinations: keyComb,
-        })
-    } else {
-        const allCombs = getStrategiesKeyCombinations(dimensions)
-
-        for (let i = 0; i < allCombs.length; i++) {
-            keyCombList.push({
-                names: allCombs[i].names,
-                combinations: generateKeyCombinationArray(
-                    allCombs[i].combinations
-                ),
-                // Calculate the size of the sub-key
-                size: allCombs[i].combinations.reduce((acc, val) => {
-                    acc = acc * val
-                    return acc
-                }, 1),
-            })
-        }
-    }
-
-    
-   // const simulationId = generateSimulationId()
-    const keyStrategy = getKeyStrategyFromDom()
-
-    // var simulationDiv = createSimulationDiv(simulationId, {
-    //     metrics,
-    //     dimensions,
-    //     epsilon,
-    //     contributionBudget,
-    //     isUseScaling,
-    //     isGranular,
-    //     batchingFrequency,
-    //     dailyConversionCount,
-    // })
-
-    const simulation = {
-        metadata: {
-            simulationTitle: generateSimulationTitle(new Date(Date.now())),
-            simulationId: generateSimulationId(),
-        },
-        inputParameters: {
-            // Used later for display
-            dailyConversionCount,
-            dimensions,
-            epsilon,
-            keyStrategy,
-            metrics,
-            batchingFrequency,
-            isUseScaling,
-        },
-        summaryReports: [],
-    }
-
-    metrics.forEach((element) => {
-        for (let i = 0; i < keyCombList.length; i++) {
-            simulation.summaryReports.push(
-            simulatePerMetric(
-                keyCombList[i],
-                element,
-                epsilon,
-                contributionBudget,
-                isUseScaling,
-                batchingFrequency,
-                getIsKeyStrategyGranularFromDom()
-                    ? getDailyValue()
-                    : Math.floor(dailyConversionCount / keyCombList[i].size),
-                i
-            ))
-        }
-
-
-
-    })
-
-    console.log(simulation);
-    displaySimulationResults_unified(simulation, "advanced")
 }
 
 window.simulateAndDisplayResultsAdvancedMode =
