+++ public/js/simple-mode.js
@@ -16,26 +16,14 @@ import {
     displayTabularData,
     initializeDisplayGeneric,
     getBatchingFrequencyFromDom,
-    getDailyConversionCountFromDom,
     getEpsilonFromDom,
-    getKeyStrategyFromDom,
-    displaySimulationResults_simpleMode,
     getIsUseScalingFromDom,
     getBudgetValueForMetricIdFromDom,
     loadPython,
+    displaySimulationResults_unified,
+    getIsKeyStrategyGranularFromDom
 } from './dom'
 import {
-    getScalingFactorForMetric,
-    getRandomLaplacianNoise,
-    calculateNoisePercentage,
-    calculateAverageNoisePercentage,
-    generateKeyCombinationArray,
-    getNoise_Rmsre,
-    generateSummaryValue,
-} from './utils.noise'
-import {
-    generateSimulationId,
-    generateSimulationTitle,
     downloadAll,
     tempSaveTable,
 } from './utils.misc'
@@ -44,6 +32,7 @@ import {
     DEFAULT_MEASUREMENT_GOALS,
     DEFAULT_DIMENSIONS,
 } from './config'
+import { triggerSimulation } from './main'
 
 let allSimulationDataTables_simpleMode = {}
 
@@ -75,172 +64,28 @@ export function downloadAll_simpleMode() {
 }
 
 export function simulateAndDisplayResultsSimpleMode() {
-    const simulation = simulate(
-        getDailyConversionCountFromDom(),
-        DEFAULT_DIMENSIONS,
-        getEpsilonFromDom(),
-        getKeyStrategyFromDom(),
+
+    const dimensionNames = DEFAULT_DIMENSIONS.map((dim) => dim.name)
+    console.log(dimensionNames)
+    const dimensionSizes = DEFAULT_DIMENSIONS.map((dim) => dim.size)
+    console.log(dimensionSizes)
+
+    const simulation = triggerSimulation(
         DEFAULT_MEASUREMENT_GOALS,
-        getBatchingFrequencyFromDom(),
+        DEFAULT_DIMENSIONS,
+        dimensionNames,
+        dimensionSizes,
+        getEpsilonFromDom(),
         CONTRIBUTION_BUDGET,
-        getIsUseScalingFromDom()
-    )
-    displaySimulationResults_simpleMode(simulation)
-}
-
-function getNumberOfDistinctKeyValuesPerKey(dimensions) {
-    return dimensions
-        .map((dimension) => dimension.size)
-        .reduce((prev, current) => prev * current, 1)
-}
-
-function simulate(
-    dailyConversionCount,
-    dimensions,
-    epsilon,
-    keyStrategy,
-    metrics,
-    batchingFrequency,
-    budget,
-    isUseScaling
-) {
-    const simulation = {
-        metadata: {
-            simulationTitle: generateSimulationTitle(new Date(Date.now())),
-            simulationId: generateSimulationId(),
-        },
-        inputParameters: {
-            // Used later for display
-            dailyConversionCount,
-            dimensions,
-            epsilon,
-            keyStrategy,
-            metrics,
-            batchingFrequency,
-            isUseScaling,
-        },
-        summaryReports: [],
-    }
-    const numberOfMetrics = metrics.length
-    for (let i = 0; i < numberOfMetrics; i++) {
-        const metric = metrics[i]
-        const isPercentage = true
-        const percentage = getBudgetValueForMetricIdFromDom(i + 1)
-        const scalingFactorForThisMetric = isUseScaling
-            ? getScalingFactorForMetric(
-                  metric,
-                  percentage,
-                  isPercentage,
-                  budget
-              )
-            : 1
-        const dailyReport_noiseless = generateNoiselessReport(
-            metric,
-            dailyConversionCount,
-            batchingFrequency,
-            dimensions,
-            scalingFactorForThisMetric
-        )
-        const dailyReport_noisy = generateNoisyReportFromNoiselessReport(
-            dailyReport_noiseless,
-            budget,
-            epsilon,
-            scalingFactorForThisMetric
-        )
-
-        const noise_ape = calculateAverageNoisePercentage(dailyReport_noisy)
-        const noise_rmsre = dailyReport_noisy.noise_rmsre
-
-        simulation.summaryReports.push({
-            data: dailyReport_noisy,
-            scalingFactor: scalingFactorForThisMetric,
-            measurementGoal: metric.name,
-            dimensionsString: dimensions.map((d) => d.name).join(' x '),
-            noiseMetrics: {
-                noise_ape_percent: Number.parseFloat(
-                    (noise_ape * 100).toFixed(3)
-                ),
-                noise_rmsre: noise_rmsre,
-            },
-        })
-    }
-    console.log(simulation)
-    return simulation
-}
-
-function generateNoiselessReport(
-    metric,
-    dailyConversionCount,
-    batchingFrequency,
-    dimensions,
-    scalingFactorForThisMetric
-) {
-    const keyCombinations = generateKeyCombinationArray(
-        dimensions.map((dim) => dim.size)
+        getIsUseScalingFromDom(),
+        getIsKeyStrategyGranularFromDom(),
+        getBatchingFrequencyFromDom(),
+        getBatchingFrequencyFromDom()
     )
 
-    // TODO fix - right now we're using j to add a deterministic variation to the numbers, so that they remain the same across simulations. It does the job but is clunky.
-
-    const report = []
-    keyCombinations.forEach((k, idx) => {
-        var summaryValue = generateSummaryValue(
-            metric,
-            idx,
-            dailyConversionCount,
-            batchingFrequency,
-            0
-        )
-
-        const summaryValue_scaled_noiseless =
-            summaryValue * scalingFactorForThisMetric
-
-        report.push({
-            // TODO, though the exact key doesn't really matter
-            key: k,
-            summaryValue,
-            summaryValue_scaled_noiseless,
-        })
-    })
-    return report
+    displaySimulationResults_unified(simulation, "simple")
 }
 
-function generateNoisyReportFromNoiselessReport(
-    noiselessReport,
-    budget,
-    epsilon,
-    scalingFactor
-) {
-    const noisyReport = noiselessReport.map((entry) => {
-        const { key, summaryValue, summaryValue_scaled_noiseless } = entry
-        const noise = getRandomLaplacianNoise(budget, epsilon)
-        return {
-            key,
-            summaryValue,
-            summaryValue_scaled_noiseless,
-            noise,
-            summaryValue_scaled_noisy: summaryValue_scaled_noiseless + noise,
-            noise_ape_individual: calculateNoisePercentage(
-                noise,
-                summaryValue_scaled_noiseless
-            ),
-        }
-    })
-
-    const allSummaryValuesPreNoise = Object.values(noisyReport).map(
-        (v) => v.summaryValue_scaled_noiseless
-    )
-    const allSummaryValuesPostNoise = Object.values(noisyReport).map(
-        (v) => v.summaryValue_scaled_noisy
-    )
-
-    noisyReport.noise_rmsre = getNoise_Rmsre(
-        allSummaryValuesPostNoise,
-        allSummaryValuesPreNoise,
-        scalingFactor
-    )
-
-    return noisyReport
-}
 
 window.downloadAll_simpleMode = downloadAll_simpleMode
 window.simulateAndDisplayResultsSimpleMode = simulateAndDisplayResultsSimpleMode
